requirements.txt
google-genai
google-adk
simple-salesforce
python-dotenv
flask
